[td] Refactor Tornado server
