+++ mage_ai/server/server.py
@@ -1,0 +1,63 @@
+from jupyter_client import KernelManager
+from subscriber import get_messages
+from websocket import WebSocketServer
+import asyncio
+import json
+import os
+import tornado.ioloop
+import tornado.web
+
+
+class ApiHandler(tornado.web.RequestHandler):
+    def check_origin(self, origin):
+        return True
+
+    def set_default_headers(self):
+        self.set_header('Content-Type', 'application/json')
+
+    def get(self):
+        value = self.get_argument('key')
+        r = json.dumps(dict(
+            method='get',
+            value=value,
+        ))
+        self.write(r)
+
+    def post(self):
+        value = self.get_argument('key')
+        r = json.dumps(dict(
+            method='post',
+            value=value,
+        ))
+        self.write(r)
+
+
+def make_app():
+    return tornado.web.Application(
+        [
+            (r'/websocket/', WebSocketServer),
+            (r'/', ApiHandler),
+        ],
+    )
+
+
+async def main():
+    manager = KernelManager()
+    manager.start_kernel()
+
+    connection_file = manager.connection_file
+    os.environ['CONNECTION_FILE'] = connection_file
+
+    app = make_app()
+    app.listen(8888)
+
+    get_messages(
+        manager.client(),
+        lambda message: WebSocketServer.send_message(message),
+    )
+
+    await asyncio.Event().wait()
+
+
+if __name__ == '__main__':
+    asyncio.run(main())
