+++ requirements.txt
@@ -3,6 +3,7 @@ Flask==2.1.2
 Jinja2==3.1.2
 MarkupSafe==2.1.1
 Werkzeug==2.1.2
+asyncio==3.4.3
 beautifulsoup4
 click==8.1.3
 flask-cors==3.0.10
@@ -10,6 +11,7 @@ importlib-metadata==4.11.3
 ipython
 itsdangerous==2.1.2
 joblib>=1.1.0
+jupyter-client==7.3.1
 jupyter-server-proxy==3.2.1
 numpyencoder==0.3.0
 numpy~=1.21.0
@@ -21,4 +23,5 @@ requests==2.28.0
 scikit-learn>=1.0
 simplejson
 six>=1.15.0
+tornado==6.1
 zipp==3.8.0
